requirements.txt
opencv-python
pillow
pyzbar